fix: styling menu bar
